Transform precompiled plugins sample to publishing convention plugins sample
Focus on the multi-repo use case and demonstrate how convention plugins can be published and consumed from another project.
