+++ subprojects/docs/docs.gradle
@@ -260,24 +260,8 @@ samples {
 
         projectInfoPlugin
 
-        precompiledScriptPluginGroovy {
-            target = "plugin"
-        }
-        precompiledScriptPluginGroovyInBuildSrc {
-            sourceDirectory = precompiledScriptPluginGroovy.sourceDirectory
-            target = "buildSrc"
-        }
-
-        precompiledScriptPluginKotlin {
-            target = "plugin"
-        }
-        precompiledScriptPluginKotlinInBuildSrc {
-            sourceDirectory = precompiledScriptPluginKotlin.sourceDirectory
-            target = "buildSrc"
-        }
-
         precompiledScriptPluginUtils {
-            target = "plugin"
+            target = "convention-plugins"
         }
         precompiledScriptPluginUtilsInBuildSrc {
             sourceDirectory = precompiledScriptPluginUtils.sourceDirectory
@@ -525,21 +509,6 @@ samples {
             }
         }
 
-        precompiledScriptPlugin {
-            sampleDirectory = samplesRoot.dir("build-organization/precompiled-script-plugin")
-            description = "Use the Groovy or Kotlin DSL in a Gradle plugin."
-            category = "Build organization"
-            common {
-                from(templates.precompiledScriptPluginUtils)
-            }
-            groovy {
-                from(templates.precompiledScriptPluginGroovy)
-            }
-            kotlin {
-                from(templates.precompiledScriptPluginKotlin)
-            }
-        }
-
         conventionPlugins {
             sampleDirectory = samplesRoot.dir("build-organization/multi-project-with-convention-plugins")
             displayName = "Sharing configuration between subprojects"
@@ -547,6 +516,16 @@ samples {
             category = "Build organization"
             common {
                 from(templates.precompiledScriptPluginUtilsInBuildSrc)
+            }
+        }
+
+        publishingConventionPlugins {
+            sampleDirectory = samplesRoot.dir("build-organization/publishing-convention-plugins")
+            displayName = "Sharing configuration between isolated projects"
+            description = "Organize and publish build configuration for reuse in other projects"
+            category = "Build organization"
+            common {
+                from(templates.precompiledScriptPluginUtils)
             }
         }
