--- subprojects/docs/src/samples/build-organization/precompiled-script-plugin/groovy/consumer-service/build.gradle
@@ -1,3 +1,0 @@
-plugins {
-    id 'myproject.service-conventions' version '1.0'
-}
