--- subprojects/docs/src/samples/build-organization/precompiled-script-plugin/kotlin/consumer-service/README.md
@@ -1,4 +1,0 @@
-## Service API
-This is a sample
-
-Other documentation would appear here.
