--- subprojects/docs/src/samples/build-organization/precompiled-script-plugin/kotlin/consumer-service/build.gradle.kts
@@ -1,3 +1,0 @@
-plugins {
-    id("myproject.service-conventions") version "1.0"
-}
