--- subprojects/docs/src/samples/build-organization/precompiled-script-plugin/kotlin/settings.gradle.kts
@@ -1,5 +1,0 @@
-rootProject.name = "consumer-service"
-
-include("consumer-service")
-
-includeBuild("plugin")
