--- subprojects/docs/src/samples/build-organization/precompiled-script-plugin/tests/build.sample.conf
@@ -1,5 +1,0 @@
-commands: [{
-    executable: gradle
-    args: build
-    flags: "--gradle-user-home=local-gradle-user-home"
-}]
