+++ subprojects/docs/src/samples/build-organization/publishing-convention-plugins/groovy/convention-plugins/build.gradle
@@ -1,0 +1,38 @@
+// tag::apply[]
+// tag::publish[]
+plugins {
+    id 'groovy-gradle-plugin'
+// end::apply[]
+    id 'maven-publish'
+// tag::apply[]
+}
+// end::apply[]
+
+group = 'com.example.conventions'
+version = '1.0'
+
+publishing {
+    repositories {
+        maven {
+            // change to point to your repo, e.g. http://my.org/repo
+            url = "$buildDir/repo"
+        }
+    }
+}
+
+tasks.named('publish') {
+    dependsOn('check')
+}
+
+// end::publish[]
+
+// tag::repositories-and-dependencies[]
+repositories {
+    gradlePluginPortal() // so that external plugins can be resolved in dependencies section
+}
+
+dependencies {
+    implementation 'gradle.plugin.com.github.spotbugs.snom:spotbugs-gradle-plugin:4.0.5'
+    testImplementation 'org.spockframework:spock-core:1.3-groovy-2.5'
+}
+// end::repositories-and-dependencies[]
