+++ subprojects/docs/src/samples/build-organization/publishing-convention-plugins/groovy/convention-plugins/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'my-org-conventions'
