+++ subprojects/docs/src/samples/build-organization/publishing-convention-plugins/groovy/convention-plugins/src/main/groovy/com.myorg.java-conventions.gradle
