+++ subprojects/docs/src/samples/build-organization/publishing-convention-plugins/groovy/convention-plugins/src/main/groovy/com.myorg.service-conventions.gradle
@@ -1,0 +1,41 @@
+// Define conventions for service projects this organization.
+// Service projects need to use the organization's Java conventions and pass some additional checks
+
+// tag::plugins[]
+plugins {
+    id 'com.myorg.java-conventions'
+}
+// end::plugins[]
+
+// The organization requires integration tests
+sourceSets {
+    integrationTest {
+        compileClasspath += sourceSets.main.output + sourceSets.test.output
+        runtimeClasspath += sourceSets.main.output + sourceSets.test.output
+    }
+}
+
+def integrationTest = tasks.register('integrationTest', Test) {
+    shouldRunAfter(test)
+
+    testClassesDirs = sourceSets.integrationTest.output.classesDirs
+    classpath = sourceSets.integrationTest.runtimeClasspath
+}
+
+configurations {
+    integrationTestImplementation.extendsFrom testImplementation
+    integrationTestRuntimeOnly.extendsFrom testRuntimeOnly
+}
+
+// The organization requires additional documentation in the README for this project
+// tag::use-java-class[]
+def readmeCheck = tasks.register('readmeCheck', com.example.ReadmeVerificationTask) {
+    // Expect the README in the project directory
+    readme = layout.projectDirectory.file("README.md")
+    // README must contain a Service API header
+    readmePatterns = ['^## Service API$']
+}
+// end::use-java-class[]
+
+tasks.named('check') { dependsOn(integrationTest, readmeCheck) }
+
