+++ subprojects/docs/src/samples/build-organization/publishing-convention-plugins/groovy/settings.gradle
@@ -1,0 +1,3 @@
+rootProject.name = 'convention-plugins'
+
+include 'convention-plugins'
