+++ subprojects/docs/src/samples/build-organization/publishing-convention-plugins/kotlin/convention-plugins/settings.gradle.kts
