+++ subprojects/docs/src/samples/build-organization/publishing-convention-plugins/kotlin/convention-plugins/src/main/kotlin/com.myorg.java-conventions.gradle.kts
