+++ subprojects/docs/src/samples/build-organization/publishing-convention-plugins/tests/publish.sample.conf
@@ -1,0 +1,5 @@
+commands: [{
+    executable: gradle
+    args: publish
+    flags: "--gradle-user-home=local-gradle-user-home"
+}]
