+++ subprojects/docs/src/samples/build-organization/publishing-convention-plugins/tests/sanityCheck.sample.conf
