--- subprojects/docs/src/samples/templates/precompiled-script-plugin-groovy/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name='my-org-conventions'
